testing

+++ src/vsr/models/VRT/modules/trunc_norm.py
@@ -34,4 +34,4 @@ def _no_grad_trunc_normal_(tensor, mean, std, a, b):
 
 
 def trunc_normal_(tensor, mean=0., std=1., a=-2., b=2.):
-    return _no_grad_trunc_normal_(tensor, mean, std, a, b)
+    return _no_grad_trunc_normal_(tensor, mean, std, a, b).type_as(tensor)
